+++ packages/react/src/index.ts
@@ -1,3 +1,4 @@
+/** @module React */
 export { default as useAnimationFrame } from './hooks/useAnimationFrame';
 export { default as useAsyncEffect } from './hooks/useAsyncEffect';
 export { default as useBindInput } from './hooks/useBindInput';
@@ -9,7 +10,10 @@ export { default as useThrottle } from './hooks/useThrottle';
 export { default as useUpload } from './hooks/useUpload';
 export { default as useLocalStorage } from './hooks/useLocalStorage';
 export { default as useWindowSize } from './hooks/useWindowSize';
-export { default as LoggingClick } from './LoggingClick';
+export { default as LoggingClick } from './components/LoggingClick';
+export { default as LoggingEvent } from './components/LoggingEvent';
+export { default as LoggingImpression } from './components/LoggingImpression';
+export { default as ImpressionArea } from './components/ImpressionArea';
 export { default as useDidMount } from './hooks/useDidMount';
 export { default as useWillUnmount } from './hooks/useWillUnmount';
 export { default as useDidUpdate } from './hooks/useDidUpdate';
@@ -18,3 +22,4 @@ export { default as useInfiniteScroll } from './hooks/useInfiniteScroll';
 export { default as useInterval } from './hooks/useInterval';
 export { default as useCombinedRefs } from './hooks/useCombinedRef';
 export { default as useScrollEvent } from './hooks/useScrollEvent';
+export { default as useClipboardCopy } from './hooks/useClipboardCopy';