+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
+/** @module Utils */
 export { default as delay } from './delay';
 export {
   createLocalStorageChangeEvent,
@@ -15,3 +16,5 @@ export { default as getObjectKeys } from './getObjectKeys';
 export * from './is';
 export * from './queryString';
 export * from './promise';
+export * from './date';
+export * from './array';