+++ packages/utils/src/promise/index.ts
@@ -7,6 +7,25 @@ export interface Defer<T> {
   reject: PromiseRejecter;
 }
 
+/**
+ * Promise defer 패턴을 사용할 수 있는 함수입니다.
+ *
+ * @example
+ * const {
+ *   promise: sdkLoadPromise,
+ *   resolve: sdkLoadPromiseResolver
+ * } = defer();
+ *
+ * async function call() {
+ *   const isLoaded = await sdkLoadPromise;
+ *   console.log('SDK 로딩 완료');
+ * }
+ *
+ * function completeLoading () {
+ *   // 로딩 중....
+ *   sdkLoadPromiseResolver(true);
+ * }
+ */
 export function defer<T>(): Defer<T> {
   let resolver: PromiseResolver<T> | null = null;
   let rejecter: PromiseRejecter | null = null;
@@ -21,4 +40,20 @@ export function defer<T>(): Defer<T> {
     resolve: resolver!,
     reject: rejecter!,
   };
+}
+
+/**
+ * Promise를 사용하여 일정 시간 동안 라인의 실행을 멈춥니다. delay의 단위로는 ms를 사용합니다.
+ *
+ * @example
+ * console.log('슬립 시작');
+ * await sleep(1000);
+ * console.log('1초 지남');
+ */
+export function sleep(delay: number) {
+  return new Promise<void>((resolve) => {
+    setTimeout(() => {
+      resolve();
+    }, delay);
+  });
 }